refactor: `NetworkExceptionHandler`를 core/handler 패키지로 이동
