+++ app/src/main/java/woowacourse/shopping/view/detail/DetailActivity.kt
@@ -14,9 +14,9 @@ import androidx.databinding.DataBindingUtil
 import woowacourse.shopping.App
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityDetailBinding
-import woowacourse.shopping.view.NetworkExceptionHandler
 import woowacourse.shopping.view.cart.CartActivity
 import woowacourse.shopping.view.core.ext.showToast
+import woowacourse.shopping.view.core.handler.NetworkExceptionHandler
 import woowacourse.shopping.view.detail.vm.DetailViewModel
 import woowacourse.shopping.view.detail.vm.DetailViewModelFactory
