+++ app/src/main/java/woowacourse/shopping/view/main/MainActivity.kt
@@ -15,9 +15,9 @@ import androidx.recyclerview.widget.GridLayoutManager
 import woowacourse.shopping.App
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityMainBinding
-import woowacourse.shopping.view.NetworkExceptionHandler
 import woowacourse.shopping.view.cart.CartActivity
 import woowacourse.shopping.view.core.ext.showToast
+import woowacourse.shopping.view.core.handler.NetworkExceptionHandler
 import woowacourse.shopping.view.detail.DetailActivity
 import woowacourse.shopping.view.main.adapter.ProductAdapter
 import woowacourse.shopping.view.main.adapter.ProductRvItems
